fix(coding-agent): clean up todo17 reconnect retry tests
