+++ .omo/evidence/task-17-quality-cleanup-diff-check.log
@@ -1,0 +1,2 @@
+/Users/yeongyu/.bash_profile: line 3: /usr/local/bin/gog.bash: No such file or directory
+COMMAND: git diff --check HEAD~2
