+++ .omo/evidence/task-17-quality-cleanup-impacted.log
@@ -1,0 +1,11 @@
+/Users/yeongyu/.bash_profile: line 3: /usr/local/bin/gog.bash: No such file or directory
+COMMAND: cd packages/coding-agent && npx tsx ../../node_modules/vitest/dist/cli.js --run test/mcp/connection.test.ts test/mcp/ping-on-call.test.ts test/mcp/idle.test.ts test/mcp/commands.test.ts test/mcp/service-lifecycle.test.ts test/mcp/register-call.test.ts test/mcp/reconnect.test.ts
+
+ RUN  v4.1.9 /Volumes/mengmotaStorage/local-workspaces/senpi-wt/senpi-mcp-plugin-w2/packages/coding-agent
+
+··················································
+
+ Test Files  7 passed (7)
+      Tests  50 passed (50)
+   Start at  05:10:04
+   Duration  7.17s (transform 5.35s, setup 81ms, import 10.05s, tests 21.36s, environment 0ms)
