+++ .omo/evidence/task-17-quality-cleanup-no-excuse.log
@@ -1,0 +1,3 @@
+/Users/yeongyu/.bash_profile: line 3: /usr/local/bin/gog.bash: No such file or directory
+COMMAND: NODE_PATH="/Volumes/mengmotaStorage/local-workspaces/senpi-wt/senpi-mcp-plugin-w2/node_modules" npx tsx /Users/yeongyu/.agents/skills/programming/scripts/typescript/check-no-excuse-rules.ts packages/coding-agent/src/core/extensions/builtin/mcp/commands.ts packages/coding-agent/src/core/extensions/builtin/mcp/expose/register.ts packages/coding-agent/src/core/extensions/builtin/mcp/health.ts packages/coding-agent/src/core/extensions/builtin/mcp/idle.ts packages/coding-agent/src/core/extensions/builtin/mcp/reconnect.ts packages/coding-agent/src/core/extensions/builtin/mcp/service-exposure.ts packages/coding-agent/src/core/extensions/builtin/mcp/service-register.ts packages/coding-agent/src/core/extensions/builtin/mcp/service.ts packages/coding-agent/test/mcp/commands.test.ts packages/coding-agent/test/mcp/fixtures/commands.ts packages/coding-agent/test/mcp/fixtures/options.ts packages/coding-agent/test/mcp/fixtures/sdk-server.ts packages/coding-agent/test/mcp/fixtures/reconnect.ts packages/coding-agent/test/mcp/reconnect.test.ts
+No violations in 14 file(s).
