+++ .omo/evidence/task-17-quality-cleanup-reconnect.log
@@ -1,10 +1,11 @@
+/Users/yeongyu/.bash_profile: line 3: /usr/local/bin/gog.bash: No such file or directory
 COMMAND: cd packages/coding-agent && npx tsx ../../node_modules/vitest/dist/cli.js --run test/mcp/reconnect.test.ts
 
  RUN  v4.1.9 /Volumes/mengmotaStorage/local-workspaces/senpi-wt/senpi-mcp-plugin-w2/packages/coding-agent
 
-······
+········
 
  Test Files  1 passed (1)
-      Tests  6 passed (6)
-   Start at  04:26:44
-   Duration  6.31s (transform 1.44s, setup 12ms, import 2.28s, tests 3.93s, environment 0ms)
+      Tests  8 passed (8)
+   Start at  05:09:49
+   Duration  6.81s (transform 1.05s, setup 10ms, import 1.69s, tests 5.02s, environment 0ms)
