+++ .omo/evidence/task-17-retry-fix-green-impacted.log
@@ -7,4 +7,3 @@
       Tests  50 passed (50)
    Start at  04:48:40
    Duration  7.55s (transform 6.13s, setup 86ms, import 11.04s, tests 21.49s, environment 0ms)
-
