+++ .omo/evidence/task-17-retry-fix-green-reconnect.log
@@ -7,4 +7,3 @@
       Tests  8 passed (8)
    Start at  04:48:40
    Duration  7.64s (transform 1.56s, setup 12ms, import 2.42s, tests 5.12s, environment 0ms)
-
