+++ .omo/evidence/task-17-senpi-mcp-plugin-red.log
@@ -50,4 +50,3 @@ Error: condition timed out
       Tests  3 failed | 2 passed (5)
    Start at  04:16:47
    Duration  18.91s (transform 1.08s, setup 11ms, import 1.72s, tests 17.09s, environment 0ms)
-
